requirements.txt
tensorflow
numpy
opencv-python-headless
patchify
matplotlib
keras
pillow
